Refactor component structure and styles: reorganize imports, restore CardSearch, Home, and NavBar components with updated styles
