+++ client/src/App.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import NavBar from "./components/NavBar";
-import CardSearch from "./components/CardSearch";
-import Home from "./components/Home"; 
+import NavBar from "./components/NavBar/NavBar";
+import CardSearch from "./components/CardSearch/CardSearch";
+import Home from "./components/Home/Home";
 
 const App = () => {
   return (
